Split CellarRegistry api into separate classes
BoozeRegistry handles everything related to booze
BrewRegistry, handles everything to do with brewing
PressingRegistry handles everthing to do with the fruit press

One major change is the removal of the meta* lists, these have been unified into
the main list by providing non-meta items with a NO_META meta (since all metadata is assumed to be greater than 0, NO_META is -1)
This removes the need to maintain 2 lists
